pyproject.toml
[project]
name = "gepa-catc"
version = "0.1.0"
description = ""
authors = [
    {name = "joshinshaji16@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "datasets (>=4.0.0,<5.0.0)",
    "dspy (>=3.0.1,<4.0.0)",
    "numpy (>=2,<2.3.0)",
    "opencv-python (>=4.12.0.88)",
    "dlib (>=20.0.0,<21.0.0)",
    "einops (>=0.8.1,<0.9.0)",
    "h5py (>=3.14.0,<4.0.0)",
    "joblib (>=1.5.1,<2.0.0)",
    "lmdb (>=1.7.3,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "lpips (>=0.1.4,<0.2.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "munch (>=4.0.0,<5.0.0)",
    "natsort (>=8.4.0,<9.0.0)",
    "networkx (>=3.5,<4.0)",
    "packaging (>=25.0,<26.0)",
    "peft (>=0.17.0,<0.18.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "scipy (>=1.16.1,<2.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "torchvision (>=0.23.0,<0.24.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "transformers (>=4.55.2,<5.0.0)",
    "wandb (>=0.21.1,<0.22.0)",
    "bert-score (>=0.3.13,<0.4.0)",
    "evaluate (>=0.4.5,<0.5.0)",
    "timm (>=1.0.19,<2.0.0)",
    "pydantic (>=2.7.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai (>=1.51.0,<1.100.0)",
    "litellm (>=1.51.9,<2.0.0)"
]

[tool.poetry]
name = "gepa_catp"
packages = [{ include = "src", from = "catp_base" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
